fix(platform): update dependencies and improve accessibility
