+++ apps/platform/app/api/auth/[...nextauth]/route.ts
@@ -1,36 +1,5 @@
 import NextAuth from 'next-auth';
-import { PrismaAdapter } from '@next-auth/prisma-adapter';
-import type { NextAuthOptions } from 'next-auth';
-import CredentialsProvider from 'next-auth/providers/credentials';
-import { prisma } from '@/utils/db';
-import { compare } from 'bcryptjs';
-
-export const authOptions: NextAuthOptions = {
-  adapter: PrismaAdapter(prisma),
-  pages: {
-    signIn: '/sign-in',
-  },
-  providers: [
-    CredentialsProvider({
-      name: 'Credentials',
-      credentials: {
-        email: { label: 'Email', type: 'text' },
-        password: { label: 'Password', type: 'password' },
-      },
-      async authorize(credentials) {
-        if (!credentials?.email || !credentials.password) return null;
-        const user = await prisma.user.findUnique({ where: { email: credentials.email } });
-        if (!user || !user.password) return null;
-        const isValid = await compare(credentials.password, user.password);
-        if (!isValid) return null;
-        return { id: user.id, email: user.email, name: user.name };
-      },
-    }),
-  ],
-  session: {
-    strategy: 'jwt',
-  },
-};
+import { authOptions } from './auth';
 
 const handler = NextAuth(authOptions);
 export { handler as GET, handler as POST };
