+++ apps/platform/app/error.tsx
@@ -15,7 +15,7 @@ export default function GlobalError({
   }, [error]);
 
   return (
-    <html>
+    <html lang="en">
       <body>
         <ErrorMessage error={error} reset={reset} />
       </body>
