+++ apps/platform/package.json
@@ -60,6 +60,7 @@
     "@testing-library/dom": "^10.4.0",
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^16.1.0",
+    "@types/bcryptjs": "^3.0.0",
     "@types/dompurify": "^3.0.0",
     "@types/lodash": "^4.17.7",
     "@types/node": "^22.14.0",
