+++ apps/platform/types/bcryptjs.d.ts
@@ -1,0 +1,8 @@
+declare module 'bcryptjs' {
+  export function hash(data: string, saltOrRounds: string | number): Promise<string>;
+  export function hashSync(data: string, saltOrRounds: string | number): string;
+  export function compare(data: string, encrypted: string): Promise<boolean>;
+  export function compareSync(data: string, encrypted: string): boolean;
+  export function genSalt(rounds?: number): Promise<string>;
+  export function genSaltSync(rounds?: number): string;
+}
